backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-dotenv==1.0.0
vk-api==11.9.9
apscheduler==3.10.4
requests==2.31.0
pillow==10.1.0
imagehash==4.3.1
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4
python-multipart==0.0.6
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
httpx==0.25.2 